+++ packages/kernel/src/worker.ts
@@ -173,8 +173,6 @@ async function loadPyodideAndPackages() {
   await pyodide.loadPackage("micropip");
   const micropip = pyodide.pyimport("micropip");
   if (wheels) {
-    await micropip.install("altair<5.2.0", { keep_going: true }); // Altair>=5.2.0 checks PyArrow version and emits an error (https://github.com/altair-viz/altair/pull/3160)
-
     console.debug(
       "Installing the wheels:",
       wheels,
@@ -287,8 +285,9 @@ async function loadPyodideAndPackages() {
   console.debug("Booting up the Streamlit server");
   // The following Python code is based on streamlit.web.cli.main_run().
   self.__streamlitFlagOptions__ = {
-    ...streamlitConfig,
+    // gatherUsageStats is disabled as default, but can be enabled explicitly by setting it to true.
     "browser.gatherUsageStats": false,
+    ...streamlitConfig,
     "runner.fastReruns": false, // Fast reruns do not work well with the async script runner of stlite. See https://github.com/whitphx/stlite/pull/550#issuecomment-1505485865.
   };
   await pyodide.runPythonAsync(`